==> CODE/4.Consine_Content_base.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import linear_kernel, cosine_similarity
from underthesea import word_tokenize, pos_tag, sent_tokenize
import warnings
import re

from tqdm.auto import tqdm
from scipy.sparse import csr_matrix, save_npz, load_npz

import pickle

%load_ext autoreload
%autoreload 2
from project_transformer import Data_Wrangling

warnings.filterwarnings("ignore")

# Load Some Data to preprocess Vietnamese languages

# EMOJI
with open('../DATA/files/emojicon.txt', 'r', encoding="utf8") as file:
    emoji_lst = file.read().split('\n')
    emoji_dict = {}
    for line in emoji_lst:
        key, value = line.split('\t')
        emoji_dict[key] = str(value)

# TEEN CODE
with open('../DATA/files/teencode.txt', 'r', encoding="utf8") as file:
    teen_lst = file.read().split('\n')
    teen_dict = {}
    for line in teen_lst:
        key, value = line.split('\t')
        teen_dict[key] = str(value)

# ENG VIET
with open('../DATA/files/english-vnmese.txt', 'r', encoding="utf8") as file:
    eng_lst = file.read().split('\n')
    eng_dict = {}
    for line in eng_lst:
        key, value = line.split('\t')
        eng_dict[key] = str(value)

# WRONG WORD
with open('../DATA/files/wrong-word.txt', 'r', encoding="utf8") as file:
    wrong_lst = file.read().split('\n')

# STOP WORD
with open('../DATA/files/vietnamese-stopwords.txt', 'r', encoding="utf8") as file:
    stop_lst = file.read().split('\n')

# Load file data:

use_cols = ['product_id','sub_category','price','rating','clean_desc','product_name','clean_prd_name']
clean_prd_df = pd.read_csv('../DATA/clean_products.csv',usecols=use_cols)

clean_prd_df.head()

## Check na values from clean data

drop_idx = clean_prd_df[clean_prd_df['clean_prd_name'].apply(lambda x: isinstance(x,float))].index

clean_prd_df = clean_prd_df.drop(index=drop_idx)

clean_prd_df = clean_prd_df.reset_index(drop=True)

# Cosine_similarity
- link https://scikit-learn.org/stable/modules/generated/sklearn.metrics.pairwise.cosine_similarity.html
- link https://en.wikipedia.org/wiki/Cosine_similarity

tf_model = TfidfVectorizer(analyzer='word', stop_words=stop_lst,max_features=None)
tf_model.fit(raw_documents=clean_prd_df['clean_prd_name'])

#tf_model.vocabulary_ 

tfidf_matrix = tf_model.transform(raw_documents=clean_prd_df['clean_prd_name'])

type(tfidf_matrix)

#tfidf_matrix[0]

# cosine_similarities = cosine_similarity(tfidf_matrix, 
#                                         tfidf_matrix,
#                                         dense_output=True)

#tfidf_matrix[0]

type(cosine_similarities)

cosine_similarities[0]

# với mỗi sản phẩm, lấy 3 sản phẩm tương quan nhất
# results = {}

# for idx, row in tqdm(clean_prd_df.iterrows(),'Calcualte the high correaltion',total=clean_prd_df.shape[0]):
#     similar_indices = cosine_similarities[idx].argsort()[-6:-1]
#     similar_items = [(cosine_similarities[idx][i]) for i in similar_indices]
#     similar_items = [(cosine_similarities[idx][i], clean_prd_df.index[i]) for i in similar_indices]
#     #print(similar_items[0:])
#     results[idx] = similar_items[0:]

#results[0]

## Save file

#np.save('../DATA/cosine_sim_prd.npy', cosine_similarities)

# Save the TF-IDF vectorizer using pickle
with open('../DATA/tfidf_vectorizer.pkl', 'wb') as f:
    pickle.dump(tf_model, f)

# Save the sparse matrix
save_npz('../DATA/tfidf_matrix.npz', tfidf_matrix)

## Apply Cosine Similarity

### Load model and matrix:

with open('../DATA/tfidf_vectorizer.pkl', 'rb') as f:
    loaded_tfidf_model = pickle.load(f)
type(loaded_tfidf_model)

load_tfidf_matrix = load_npz('../DATA/tfidf_matrix.npz')
type(load_tfidf_matrix)

### Build Function

final_df = clean_prd_df.reset_index().rename(columns={'index':'item_id'})
print("Shape of final DF:", final_df.shape)
final_df.head()

final_df.to_csv(path_or_buf='../DATA/final_details.csv',index=False)

text_ = final_df.loc[final_df['item_id'] == 10,'clean_prd_name'].values[0]
text_

txt_sparse_matrix = loaded_tfidf_model.transform([text_])
type(txt_sparse_matrix)

cosine_text_ = cosine_similarity(X=txt_sparse_matrix,Y=load_tfidf_matrix)
cosine_text_

top5_prd = cosine_text_.flatten().argsort()[-6:-1]
top5_score = [(cosine_text_.flatten()[i]) for i in top5_indices]

# Extract Product
ls_item = []
for id, sim in zip(top5_prd,top5_score):
    sim_item = {'ite_id': id, 'score':sim}
    ls_item.append(sim_item)

ls_item

# Sample text
text_ = final_df.loc[final_df['item_id'] == 10,'clean_prd_name'].values[0]

# Get Sparse Matrix from TFIDF
txt_sparse_matrix = loaded_tfidf_model.transform([text_])

# Calcualte the cosine score
cosine_text_ = cosine_similarity(X=txt_sparse_matrix,Y=load_tfidf_matrix)

top5_prd = cosine_text_.flatten().argsort()[-6:-1]
top5_score = [(cosine_text_.flatten()[i]) for i in top5_indices]

# Extract Product
ls_item = []
for id, sim in zip(top5_prd,top5_score):
    sim_item = {'item_id': id, 'score':sim}
    ls_item.append(sim_item)

ls_item

## Function

def recommend_cosine(text_,
                     df,
                     tf_idf_model,
                     tf_idf_arr,
                     num = 5):
    # Get Sparse Matrix from TFIDF
    txt_sparse_matrix = loaded_tfidf_model.transform([text_])
    
    # Calcualte the cosine score
    cosine_text_ = cosine_similarity(X=txt_sparse_matrix,
                                     Y=load_tfidf_matrix)
    
    top5_prd = cosine_text_.flatten().argsort()[-num:-1]
    top5_score = [(cosine_text_.flatten()[i]) for i in top5_indices]
    
    # Extract Product
    ls_item = []
    for id, sim in zip(top5_prd,top5_score):
        sim_item = {'item_id': id, 'score':sim}
        ls_item.append(sim_item)
    print(f"San Pham user nhap: {text_}\n\n")
    for rec in ls_item:
        #print(rec[1])
        print(f"Recommnended:\tItem ID: {rec['item_id']}, {df.loc[df['item_id'] == rec['item_id'],'product_name'].values[0]}, (score: {rec['score']})\n")
    get_item =  df['item_id'].isin(top5_prd)
    
    return df[get_item]

text_ = final_df.loc[final_df['item_id'] == 10,'clean_prd_name'].values[0]
#print(text_)
recm_df = recommend_cosine(text_=text_,df=final_df,tf_idf_model=loaded_tfidf_model,tf_idf_arr=load_tfidf_matrix)

recm_df
