==> CODE/5. Gensim_Content_base.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import linear_kernel, cosine_similarity
from underthesea import word_tokenize, pos_tag, sent_tokenize
import warnings
from gensim import corpora, models, similarities
#import jieba
import re

import matplotlib.pyplot as plt 
import seaborn as sns
import glob
import os 
from tqdm.auto import tqdm
import re

# Load file data:

use_cols = ['product_id','sub_category','price','rating','clean_desc','product_name','clean_prd_name']
clean_prd_df = pd.read_csv('../DATA/final_details.csv',usecols=use_cols)

clean_prd_df.head()

## Check na values from clean data

#clean_prd_df[clean_prd_df['clean_prd_name'].apply(lambda x: isinstance(x,float))]

drop_idx = clean_prd_df[clean_prd_df['clean_prd_name'].apply(lambda x: isinstance(x,float))].index
drop_idx

clean_prd_df = clean_prd_df.drop(index=drop_idx)

def export_dict_gensim(col_values: np.array):
    # Tokenize(split) the sentences into words
    #products_gem = [ for x in col_values]
    products_gem = list()
    for x in tqdm(col_values,"Tokenize data"):
        if isinstance(x,str):
            split_txt = [text for text in x.split()]
            products_gem.append(split_txt)
            
    # remove some special elements in texts
    # number
    products_gem_re = [[re.sub('[0-9]+', '', e) for e in text] for text in products_gem]
    # special symbols
    special_ls = ['', ' ', ',', '.', '...', '-', ':', ';', '?',
                  '%', '_%', '(', ')', '+', '/', 'g', 'ml']
    products_gem_re = [[t.lower() for t in text if t not in special_ls] for text in products_gem_re]

    # Obtain the number of features based on dictionary: Use corpora.Dictionary
    dictionary = corpora.Dictionary(products_gem_re)
    print('Numbers of features in **Dictionary**:',len(dictionary.token2id))

    # Obtain corpus based on dictionary (dense matrix)
    corpus = [dictionary.doc2bow(text) for text in products_gem_re]
    return dictionary, corpus

prd_name_values = clean_prd_df['clean_prd_name'].values

name_prd_dict, name_prd_corpus = export_dict_gensim(col_values=prd_name_values)

- View the sample of words from Gensim

list(name_prd_dict.token2id.keys())[0:10]

name_prd_corpus[0]

# Apply model TFIDF:

# Create the TF-IDF model
tfidf = models.TfidfModel(corpus=name_prd_corpus, smartirs='ntc')

# Show the TF-IDF weights
for doc in tfidf[name_prd_corpus][0:10]:
    print([[name_prd_dict[id], np.around(freq, decimals=2)] for id, freq in doc])
    print("\n\n")

# tính toán sự tương tự trong ma trận thưa thớt
feature_cnt = len(name_prd_dict.token2id)
index = similarities.SparseMatrixSimilarity(tfidf[name_prd_corpus],
                                            num_features = feature_cnt)

#index[tfidf[name_prd_corpus]][0]

index[tfidf[name_prd_corpus[0]]]

type(index)

# Save Data:

type(tfidf)

tfidf.save('../DATA/Gensim/tfidf_model')

name_prd_dict.save('../DATA/Gensim/corpus_dictionary')

index.save("../DATA/Gensim/similarity_matrix.index")

#data_result = index[tfidf[name_prd_corpus]]
